file_client/requirements.txt
fastapi[all]
grpcio
grpcio-tools
pytest